move zkapp_commands_to_transactions to block
